// src/app/models/user.interface.ts
export interface UserI {

    correo:string;
    estado_registro:number;
    fecha_modificacion:string;
    fecha_nacimiento:string;
    fecha_registro:string;
    id_departamento:number;
    id_pais:number;

    id_usuario: number;

    nombre_completo: string;
    nombres: string;
    numero_celular: string;
    numero_documento: string;
    primer_apellido: string;
    rol: string;
    segundo_apellido: string;
    usuario_modificacion: string;
    usuario_registro: string;
}